fix(codegen): fix
